Generate velocity config. Refactor generator.
