+++ .python-version
@@ -1,1 +1,1 @@
-3.10.5
+3.11-dev
