+++ docker-compose.web.yml
@@ -1,30 +1,22 @@
-version: '3.4'
+version: '3.5'
 
 services:
-  yakumo-dash:
-    container_name: yakumo-dash
-    image: yukkuricraft/yakumo-dash
-    environment:
-      VIRTUAL_HOST: "yakumo.yukkuricraft.net"
-      LETSENCRYPT_HOST: "yakumo.yukkuricraft.net"
-    ports:
-      - 25698:80
-
+  # For yakumo-dash, run yakumo-dash-dev from yakumo-dash repo.
   yc_api:
     container_name: yc-api
     image: yukkuricraft/yc-docker-api
     restart: 'on-failure'
     environment:
-      VIRTUAL_HOST: "mc.yukkuricraft.net"
-      LETSENCRYPT_HOST: "mc.yukkuricraft.net"
+      VIRTUAL_HOST: "api.yukkuricraft.net"
+      VIRTUAL_PORT: 5001
+      LETSENCRYPT_HOST: "api.yukkuricraft.net"
     volumes:
     - /usr/bin/docker:/usr/bin/docker
     - /usr/bin/docker-compose:/usr/bin/docker-compose
     - /var/run/docker.sock:/var/run/docker.sock
     - ./:/app
-    ports:
-    - 5000:5000
-    - 5001:5001
+    networks:
+      - yc-web-dev
 
   nginx-proxy:
     container_name: nginx-proxy
@@ -37,6 +29,8 @@ services:
       - certs:/etc/nginx/certs:ro
       - html:/usr/share/nginx/html
       - vhost:/etc/nginx/vhost.d
+    networks:
+      - yc-web-dev
 
   letsencrypt-nginx-proxy:
     container_name: letsencrypt-nginx-proxy
@@ -53,9 +47,14 @@ services:
       - acme:/etc/acme.sh
     depends_on:
       - nginx-proxy
+    networks:
+      - yc-web-dev
 
 volumes:
   certs:
   html:
   vhost:
   acme:
+
+networks:
+  yc-web-dev:
