--- src/constants.py
@@ -1,2 +1,0 @@
-DOCKER_COMPOSE_TEMPLATE_NAME = "templates/docker-compose.tpl.yml"
-CONFIG_NAME = "world_group_config.yml"
