+++ src/generator/constants.py
@@ -1,0 +1,5 @@
+DOCKER_COMPOSE_TEMPLATE_NAME = "templates/docker-compose.tpl.yml"
+
+WORLD_GROUP_CONFIG_NAME = "world_group_config.yml"
+
+VELOCITY_CONFIG_TEMPLATE_NAME = "templates/velocity.tpl.toml"
