+++ src/generator/templates/docker-compose.tpl.yml
@@ -9,7 +9,6 @@ custom_extensions:
   mc_service_template: # Config used to run all world group containers.
     image: yukkuricraft/minecraft-server
     depends_on:
-      - velocity
       - mysql
     tty: true
     stdin_open: true
@@ -39,11 +38,28 @@ custom_extensions:
         target: /plugins-bindmount-prod
         read_only: ${BINDMOUNT_RO}
       - mcdata_<<WORLDGROUP>>:/data
-      - ./secrets/configs/${ENV}/worlds/default:/yc-default-configs
-      - ./secrets/configs/${ENV}/worlds/<<WORLDGROUP>>/plugins:/yc-plugin-configs
-      - ./secrets/configs/${ENV}/worlds/<<WORLDGROUP>>/server:/yc-server-configs
+      - ${YC_REPO_ROOT}/secrets/configs/${ENV}/worlds/default:/yc-default-configs
+      - ${YC_REPO_ROOT}/secrets/configs/${ENV}/worlds/<<WORLDGROUP>>/plugins:/yc-plugin-configs
+      - ${YC_REPO_ROOT}/secrets/configs/${ENV}/worlds/<<WORLDGROUP>>/server:/yc-server-configs
       - ycworldsvolume_<<WORLDGROUP>>:/worlds-volume-dev # This volume is unused in prod. Only in dev.
       - ycpluginsvolume_<<WORLDGROUP>>:/plugins-volume-dev # Also unused in prod
+
+  velocity_template:
+    depends_on:
+      - mysql
+    image: itzg/bungeecord
+    environment:
+      DEBUG: "false"
+      TYPE: "VELOCITY"
+      SYNC_SKIP_NEWER_IN_DESTINATION: "false"
+    volumes:
+      - velocity-${ENV}:/server
+      - ${YC_REPO_ROOT}/gen/velocity-${ENV}.toml:/config/velocity.toml
+      - ${YC_REPO_ROOT}/secrets/velocity/forwarding.secret:/config/forwarding.secret
+    ports:
+      - '${VELOCITY_PORT}:25577'
+    networks:
+      - ycnet
 
 
 # "Static" services
@@ -55,24 +71,10 @@ services:
     networks:
       - ycnet
 
-  velocity:
-    depends_on:
-      - mysql
-    image: itzg/bungeecord
-    environment:
-      TYPE: "VELOCITY"
-      CFG_MOTD: "Powered by Docker"
-    volumes:
-      - velocity-${ENV}:/server
-      - ./secrets/velocity/:/config/
-    ports:
-      - '${VELOCITY_PORT}:25600'
-    networks:
-      - ycnet
-
 
 # Volumes will also be dynamically generated for each "enabled" world group.
 volumes:
 
+# Networks will be automatically generated as ycnet-${ENV}
 networks:
   ycnet:
